refactor(rationale): move parser to agent_sessions, colocate with instruction
Fix layering violation: agent_sessions was importing from swarm, creating a
cycle with existing swarm imports of agent_sessions.

Move parse_rationale, _empty, _decorated, and regex patterns from
swarm/rationale.py to agent_sessions/rationale.py. Both instruction and
parser are now colocated in agent_sessions, which is correct per #4796
issue description: "each typed-output contract owns a named instruction
block and its parser".

swarm/rationale.py now re-exports parse_rationale for backward
compatibility, so existing swarm imports continue to work unchanged while
the layering is corrected.

Update production imports:
- agent_sessions/mcp.py: agent_sessions.rationale
- agent_sessions/router.py: agent_sessions.rationale
- swarm/compare_router.py: agent_sessions.rationale (via swarm re-export)
- swarm/steps.py: agent_sessions.rationale (via swarm re-export)
- swarm/router.py: agent_sessions.rationale (via swarm re-export)
- swarm/view.py: agent_sessions.rationale (via swarm re-export)

Update BUILD deps to reflect the new structure. agent_sessions:rationale
now owns parsing; swarm:rationale depends on agent_sessions:rationale for
re-export.

No cycle: agent_sessions has no swarm imports, swarm depends on
agent_sessions (preexisting, correct).
